bash: Alias vim to 'gvim -v' for clipboard support
On Fedora, the default vim executable is not compiled with X11/clipboard
support.  'gvim -v' runs the executable from the vim-X11 package (still
in the terminal!) which is compiled with clipboard support.

Fedora also has the 'vimx' command, but this is not as portable across
various Linux distros.  The gvim variant should work wherever GVim is
installed.

Reference: https://superuser.com/a/194750