Use 'dot-' prefix for hidden files
`stow` supports the `--dotfiles` option which causes files beginning
with the prefix `dot-` in the stow directory to be symlinked with a literal
period character instead.  This avoids the need to have hidden files in
the stow directory.

Unfortunately, this option is not fully working for hidden directories
(see https://github.com/aspiers/stow/issues/33).  Hence, packages
containing hidden directories remain unchanged for now.

Add a README file to help with remembering how the installation works
for each package.