tmux: Use absolute pathname for tmux-preview-pane
The tmux-preview-pane utility is installed in "$HOME/bin" which is
placed in the $PATH variable in the .bashrc file.  However, depending on
how tmux is started, the .bashrc file may not be sourced and the
modified $PATH may not make it into the default tmux environment.

There are several solutions to this, most of which involve tricks to
force the sourcing of .bashrc and/or to manually modify tmux's default
environment, but none of these are particularly elegant.  The simplest
approach is to put the full path to the tmux-preview-pane script in the
tmux configuration file (this should be a relatively static location
given how the rcfiles repository is structured).